feat(diagram): Implement DiagramBuilder, DiagramActivationWalker
This adds helper classes for constructing activation bar parts and
message parts recursively based on the ActivationWalker concept.
The diagram parts are first added to a DiagramBuilder which ensures
correct numbering of messages among other things, then the parts arrays
are built and passed to the immutable Diagram class.
The messages and activation bars cannot yet be drawn because layout
logic for them is missing.